Split ConsumeSite in 2 classes to reduce complexity
